Rework git and finalize package impl
