+++ external-crates/move/crates/move-package-alt/src/cli/parse.rs
@@ -44,7 +44,7 @@ impl Parse {
             let lockfile_path = path.join("Move.lock");
             if manifest_path.exists() {
                 println!("Manifest file found at: {:?}", manifest_path);
-                let manifest = Manifest::<Vanilla>::read_from(&manifest_path);
+                let manifest = Manifest::<Vanilla>::read_from_file(&manifest_path);
                 match manifest {
                     Ok(manifest) => {
                         println!("{:?}", manifest);
@@ -56,7 +56,7 @@ impl Parse {
             }
             if lockfile_path.exists() {
                 println!("Lockfile found at: {:?}", lockfile_path);
-                let lockfile = Lockfile::<Vanilla>::read_from(&lockfile_path);
+                let lockfile = Lockfile::<Vanilla>::read_from_dir(&lockfile_path);
                 match lockfile {
                     Ok(lockfile) => {
                         println!("{:?}", lockfile);
@@ -69,7 +69,7 @@ impl Parse {
         }
 
         if let Some(manifest_path) = manifest {
-            let m = Manifest::<Vanilla>::read_from(manifest_path);
+            let m = Manifest::<Vanilla>::read_from_file(manifest_path);
             match m {
                 Ok(manifest) => {
                     println!("{:?}", manifest);
@@ -81,7 +81,7 @@ impl Parse {
         }
 
         if let Some(lockfile_path) = lockfile {
-            let lockfile = Lockfile::<Vanilla>::read_from(lockfile_path);
+            let lockfile = Lockfile::<Vanilla>::read_from_dir(lockfile_path);
             match lockfile {
                 Ok(lockfile) => {
                     println!("{:?}", lockfile);
