+++ external-crates/move/crates/move-package-alt/src/dependency/local.rs
@@ -16,21 +16,50 @@ use serde_spanned::Spanned;
 
 use super::PinnedDependencyInfo;
 
-// TODO: PinnedLocalDependencies should be different from UnpinnedLocalDependency - the former also
-// needs an absolute filesystem path (which doesn't get serialized to the lockfile)
 #[derive(Debug, Serialize, Deserialize)]
 #[derive_where(Clone, PartialEq)]
+#[serde(untagged)]
 #[serde(bound = "")]
-pub struct LocalDependency<F: MoveFlavor + ?Sized> {
+pub enum LocalDependency<F: MoveFlavor + ?Sized> {
+    /// A local dependency that is not pinned
+    Unpinned(UnpinnedLocalDependency),
+    /// A local dependency that is pinned, containing additional metadata about the dependency's
+    /// parent.
+    Pinned(PinnedLocalDependency<F>),
+}
+
+// TODO: PinnedLocalDependencies should be different from UnpinnedLocalDependency - the former also
+// needs an absolute filesystem path (which doesn't get serialized to the lockfile)
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
+pub struct UnpinnedLocalDependency {
     /// The path on the filesystem, relative to the location of the containing file (which is
     /// stored in the `Located` wrapper)
     local: PathBuf,
+}
 
-    parent: Option<Box<PinnedDependencyInfo<F>>>,
+#[derive(Debug, Serialize, Deserialize)]
+#[derive_where(Clone, PartialEq)]
+#[serde(bound = "")]
+pub struct PinnedLocalDependency<F: MoveFlavor + ?Sized> {
+    /// The path on the filesystem, relative to the location of the containing file (which is
+    /// stored in the `Located` wrapper)
+    local: PathBuf,
+    /// The dependency info for the package that contains this local dependency
+    parent: Box<PinnedDependencyInfo<F>>,
 }
 
 impl<F: MoveFlavor> LocalDependency<F> {
-    /// Returns the path to the local dependency
+    /// The path on the filesystem, relative to the location of the containing file
+    pub fn path(&self) -> PackageResult<PathBuf> {
+        match self {
+            LocalDependency::Unpinned(dep) => dep.path(),
+            LocalDependency::Pinned(dep) => dep.path(),
+        }
+    }
+}
+
+impl UnpinnedLocalDependency {
+    /// The path on the filesystem, relative to the location of the containing file
     pub fn path(&self) -> PackageResult<PathBuf> {
         let path = fs::canonicalize(&self.local)?;
         Ok(path)
@@ -43,8 +72,15 @@ impl<F: MoveFlavor> LocalDependency<F> {
     // }
 }
 
+impl<F: MoveFlavor> PinnedLocalDependency<F> {
+    pub fn path(&self) -> PackageResult<PathBuf> {
+        let path = fs::canonicalize(&self.local)?;
+        Ok(path)
+    }
+}
+
 // TODO: dead code
-impl<F: MoveFlavor> TryFrom<(&Path, toml_edit::Value)> for LocalDependency<F> {
+impl TryFrom<(&Path, toml_edit::Value)> for UnpinnedLocalDependency {
     type Error = anyhow::Error; // TODO
 
     fn try_from(value: (&Path, toml_edit::Value)) -> Result<Self, Self::Error> {
