+++ external-crates/move/crates/move-package-alt/src/lib.rs
@@ -11,6 +11,7 @@ pub mod cli;
 pub mod dependency;
 pub mod errors;
 pub mod flavor;
+pub mod git;
 pub mod graph;
 pub mod jsonrpc;
 pub mod package;
